Added 'TransferCards' function. Incomplete.

+++ DistributeCards.py
@@ -2,7 +2,7 @@
 def DisCards(ListOfCards):
 	# Extra safety check
 	if len(ListOfCards) % 2 == 0:
-		list1 = ListOfCards[:len(ListOfCards)/2]
-		list2 = ListOfCards[len(ListOfCards)/2:]
+		list1 = ListOfCards[:int(len(ListOfCards)/2)]
+		list2 = ListOfCards[int(len(ListOfCards)/2):]
 		return list1, list2
